+++ test/pc/setup-dkml-darwin_arm64.sh
@@ -222,6 +222,7 @@ export opam_root_cacheable="${PC_PROJECT_DIR}/.ci/o"
 export abi_pattern="macos-darwin_all-silicon"
 export no_gh="true"
 export dkml_host_abi="darwin_arm64"
+export dkml_target_abi="darwin_arm64"
 export opam_root="${PC_PROJECT_DIR}/.ci/o"
 
 
@@ -1304,6 +1305,7 @@ GIT_LOCATION=${GIT_LOCATION:-}
 Matrix
 ------
 dkml_host_abi=$dkml_host_abi
+dkml_target_abi=$dkml_target_abi
 abi_pattern=$abi_pattern
 opam_root=${opam_root}
 opam_root_cacheable=${opam_root_cacheable}
@@ -2337,6 +2339,7 @@ do_use_vsstudio() {
             echo "  [DKML_COMPILE_VS_MSVSPREFERENCE = \"$E_VS_MSVSPREFERENCE\"]" >>"$OPSC"
             echo "  [DKML_COMPILE_VS_CMAKEGENERATOR = \"$E_VS_CMAKEGENERATOR\"]" >>"$OPSC"
             echo "  [DKML_HOST_ABI = \"${dkml_host_abi}\"]" >>"$OPSC"
+            echo "  [DKML_TARGET_ABI = \"${dkml_target_abi}\"]" >>"$OPSC"
             echo ']' >>"$OPSC"
             cat "$OPSC" # print
             ;;
@@ -2350,6 +2353,7 @@ do_use_vsstudio() {
             opamrun option --switch "$do_use_vsstudio_NAME" setenv+="DKML_COMPILE_VS_MSVSPREFERENCE = \"$E_VS_MSVSPREFERENCE\""
             opamrun option --switch "$do_use_vsstudio_NAME" setenv+="DKML_COMPILE_VS_CMAKEGENERATOR = \"$E_VS_CMAKEGENERATOR\""
             opamrun option --switch "$do_use_vsstudio_NAME" setenv+="DKML_HOST_ABI = \"${dkml_host_abi}\""
+            opamrun option --switch "$do_use_vsstudio_NAME" setenv+="DKML_TARGET_ABI = \"${dkml_target_abi}\""
             opamrun option --switch "$do_use_vsstudio_NAME" setenv # print
             ;;
         esac
@@ -2471,6 +2475,7 @@ Finished setup.
 
 To continue your testing, run:
   export dkml_host_abi='${dkml_host_abi}'
+  export dkml_target_abi='${dkml_target_abi}'
   export abi_pattern='${abi_pattern}'
   export opam_root='${opam_root}'
   export exe_ext='${exe_ext:-}'