+++ test/pc/setup-dkml-linux_x86_64.sh
@@ -225,6 +225,7 @@ export dkml_host_os="linux"
 export opam_root_cacheable=".ci/o"
 export abi_pattern="manylinux2014-linux_x86_64"
 export dkml_host_abi="linux_x86_64"
+export dkml_target_abi="linux_x86_64"
 export opam_root=".ci/o"
 export dockcross_image="dockcross/manylinux2014-x64"
 export dockcross_run_extra_args="--platform linux/amd64"
@@ -1310,6 +1311,7 @@ GIT_LOCATION=${GIT_LOCATION:-}
 Matrix
 ------
 dkml_host_abi=$dkml_host_abi
+dkml_target_abi=$dkml_target_abi
 abi_pattern=$abi_pattern
 opam_root=${opam_root}
 opam_root_cacheable=${opam_root_cacheable}
@@ -2343,6 +2345,7 @@ do_use_vsstudio() {
             echo "  [DKML_COMPILE_VS_MSVSPREFERENCE = \"$E_VS_MSVSPREFERENCE\"]" >>"$OPSC"
             echo "  [DKML_COMPILE_VS_CMAKEGENERATOR = \"$E_VS_CMAKEGENERATOR\"]" >>"$OPSC"
             echo "  [DKML_HOST_ABI = \"${dkml_host_abi}\"]" >>"$OPSC"
+            echo "  [DKML_TARGET_ABI = \"${dkml_target_abi}\"]" >>"$OPSC"
             echo ']' >>"$OPSC"
             cat "$OPSC" # print
             ;;
@@ -2356,6 +2359,7 @@ do_use_vsstudio() {
             opamrun option --switch "$do_use_vsstudio_NAME" setenv+="DKML_COMPILE_VS_MSVSPREFERENCE = \"$E_VS_MSVSPREFERENCE\""
             opamrun option --switch "$do_use_vsstudio_NAME" setenv+="DKML_COMPILE_VS_CMAKEGENERATOR = \"$E_VS_CMAKEGENERATOR\""
             opamrun option --switch "$do_use_vsstudio_NAME" setenv+="DKML_HOST_ABI = \"${dkml_host_abi}\""
+            opamrun option --switch "$do_use_vsstudio_NAME" setenv+="DKML_TARGET_ABI = \"${dkml_target_abi}\""
             opamrun option --switch "$do_use_vsstudio_NAME" setenv # print
             ;;
         esac
@@ -2477,6 +2481,7 @@ Finished setup.
 
 To continue your testing, run:
   export dkml_host_abi='${dkml_host_abi}'
+  export dkml_target_abi='${dkml_target_abi}'
   export abi_pattern='${abi_pattern}'
   export opam_root='${opam_root}'
   export exe_ext='${exe_ext:-}'